Simplify generating UNIX timestamp
You are generating a `datetime` object, converting it to a timestring,
parsing that again to convert it back to a `datetime` object to then
convert it to a UNIX timestanmp. That's a very convoluted way of getting
a timestamp.

The current code:

```python
>>> int(dt.strptime(str(parse(str(dt.now()))), '%Y-%m-%d %H:%M:%S.%f').timestamp()) * 1000
1705790326000
```

is basically identical to just using the datetime object directly:

```python
>>> int(dt.now().timestamp()) * 1000
1705790327000
```

but it's even simpler to use the `time` library you are already using:

```python
>>> int(time.time()) * 1000
1705790328000
```