Introduce flak8
This patch adds a GitHub Actions workflow which runs `flake8` as style
checker and basic static code analysis tool.

For now, this still excludes the line length check. We should get to
that eventually, but most of the warnings probably get fixed when we
address #7 and introduce a logger anyway.