Install binary in container
Instead of copying the source code and executing the package, this patch
only installs the binary in the container, discarding additional source
files.

This will not make much of a difference but should be slightly faster
and result in slightly smaller containers.

This also now uses the arguably nicer location of
`/etc/camera-control.yml` for the default configuration file. This means
that we will have the same configuration location for all installation
types.